refactor: lsp and resolver code structure and cargo clippy
