+++ kclvm/tools/src/LSP/src/notification.rs
@@ -9,7 +9,7 @@ use crate::{
     from_lsp,
     state::LanguageServerState,
     util::apply_document_changes,
-    util::{build_word_index_for_file_content, word_index_add, word_index_subtract},
+    word_index::{build_word_index_with_content, word_index_add, word_index_subtract},
 };
 
 impl LanguageServerState {
@@ -85,7 +85,7 @@ impl LanguageServerState {
         let path = from_lsp::abs_path(&text_document.uri)?;
         self.log_message(format!("on did_change file: {:?}", path));
 
-        // update vfs
+        // Update vfs
         let vfs = &mut *self.vfs.write();
         let file_id = vfs
             .file_id(&path.clone().into())
@@ -96,12 +96,11 @@ impl LanguageServerState {
         apply_document_changes(&mut text, content_changes);
         vfs.set_file_contents(path.into(), Some(text.clone().into_bytes()));
 
-        // update word index
-        let old_word_index = build_word_index_for_file_content(old_text, &text_document.uri, true);
-        let new_word_index =
-            build_word_index_for_file_content(text.clone(), &text_document.uri, true);
+        // Update word index
+        let old_word_index = build_word_index_with_content(&old_text, &text_document.uri, true);
+        let new_word_index = build_word_index_with_content(&text, &text_document.uri, true);
         let binding = from_lsp::file_path_from_url(&text_document.uri)?;
-        let file_path = Path::new(&binding); //todo rename
+        let file_path = Path::new(&binding);
         let word_index_map = &mut *self.word_index_map.write();
         for (key, value) in word_index_map {
             let workspace_folder_path = Path::new(key.path());
