+++ kclvm/tools/src/LSP/src/util.rs
@@ -6,8 +6,7 @@ use kclvm_ast::ast::{
 use kclvm_ast::node_ref;
 use kclvm_ast::pos::ContainsPos;
 
-use kclvm_driver::kpm::fetch_metadata;
-use kclvm_driver::{get_kcl_files, lookup_compile_unit};
+use kclvm_driver::lookup_compile_unit;
 use kclvm_error::Diagnostic;
 use kclvm_error::Position as KCLPos;
 use kclvm_parser::entry::get_dir_files;
@@ -17,24 +16,18 @@ use kclvm_sema::core::global_state::GlobalState;
 use kclvm_sema::namer::Namer;
 
 use kclvm_sema::resolver::resolve_program_with_opts;
-use kclvm_sema::resolver::scope::{CachedScope, ProgramScope};
+use kclvm_sema::resolver::scope::{KCLScopeCache, ProgramScope};
 
-use kclvm_span::symbol::reserved;
-use kclvm_utils::pkgpath::rm_external_pkg_name;
-use lsp_types::{Location, Position, Range, Url};
-use parking_lot::{RwLock, RwLockReadGuard};
+use crate::from_lsp;
+use crate::state::KCLVfs;
+use lsp_types::Url;
+use parking_lot::RwLockReadGuard;
 use ra_ap_vfs::{FileId, Vfs};
 use serde::{de::DeserializeOwned, Serialize};
 
-use std::collections::HashMap;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
-use std::{
-    fs,
-    sync::{Arc, Mutex},
-};
-
-use crate::from_lsp;
+use std::{fs, sync::Arc};
 
 #[allow(unused)]
 /// Deserializes a `T` from a json value.
@@ -51,7 +44,8 @@ pub(crate) fn to_json<T: Serialize>(value: T) -> anyhow::Result<serde_json::Valu
     serde_json::to_value(value).map_err(|e| anyhow::anyhow!("could not serialize to json: {}", e))
 }
 
-pub fn get_file_name(vfs: RwLockReadGuard<Vfs>, file_id: FileId) -> anyhow::Result<String> {
+/// Get the filename from the given VFS using the file id.
+pub(crate) fn get_file_name(vfs: RwLockReadGuard<Vfs>, file_id: FileId) -> anyhow::Result<String> {
     if let Some(path) = vfs.file_path(file_id).as_path() {
         Ok(path
             .as_ref()
@@ -69,12 +63,12 @@ pub fn get_file_name(vfs: RwLockReadGuard<Vfs>, file_id: FileId) -> anyhow::Resu
 pub(crate) struct Param {
     pub file: String,
     pub module_cache: Option<KCLModuleCache>,
-    pub scope_cache: Option<Arc<Mutex<CachedScope>>>,
+    pub scope_cache: Option<KCLScopeCache>,
 }
 
 pub(crate) fn parse_param_and_compile(
     param: Param,
-    vfs: Option<Arc<RwLock<Vfs>>>,
+    vfs: Option<KCLVfs>,
 ) -> anyhow::Result<(Program, ProgramScope, IndexSet<Diagnostic>, GlobalState)> {
     let (mut files, opt) = lookup_compile_unit(&param.file, true);
     if !files.contains(&param.file) {
@@ -85,7 +79,7 @@ pub(crate) fn parse_param_and_compile(
     let mut opt = opt.unwrap_or_default();
     opt.load_plugins = true;
 
-    // update opt.k_code_list
+    // Update opt.k_code_list
     if let Some(vfs) = vfs {
         let mut k_code_list = load_files_code_from_vfs(&files, vfs)?;
         opt.k_code_list.append(&mut k_code_list);
@@ -105,12 +99,11 @@ pub(crate) fn parse_param_and_compile(
 
     let gs = GlobalState::default();
     let gs = Namer::find_symbols(&program, gs);
-    let node_ty_map = prog_scope.node_ty_map.clone();
-    let global_state = AdvancedResolver::resolve_program(&program, gs, node_ty_map);
+    let gs = AdvancedResolver::resolve_program(&program, gs, prog_scope.node_ty_map.clone());
 
     sess.append_diagnostic(prog_scope.handler.diagnostics.clone());
     let diags = sess.1.borrow().diagnostics.clone();
-    Ok((program, prog_scope, diags, global_state))
+    Ok((program, prog_scope, diags, gs))
 }
 
 /// Update text with TextDocumentContentChangeEvent param
@@ -131,7 +124,7 @@ pub(crate) fn apply_document_changes(
     }
 }
 
-fn load_files_code_from_vfs(files: &[&str], vfs: Arc<RwLock<Vfs>>) -> anyhow::Result<Vec<String>> {
+fn load_files_code_from_vfs(files: &[&str], vfs: KCLVfs) -> anyhow::Result<Vec<String>> {
     let mut res = vec![];
     let vfs = &mut vfs.read();
     for file in files {
@@ -727,641 +720,5 @@ fn build_identifier_from_ty_string(ty: &NodeRef<Type>, pos: &KCLPos) -> Option<N
         }
         Type::Literal(_) => None,
         Type::Function(_) => None,
-    }
-}
-
-/// [`get_real_path_from_external`] will ask for the local path for [`pkg_name`] with subdir [`pkgpath`].
-/// If the external package, whose [`pkg_name`] is 'my_package', is stored in '\user\my_package_v0.0.1'.
-/// The [`pkgpath`] is 'my_package.examples.apps'.
-///
-/// [`get_real_path_from_external`] will return '\user\my_package_v0.0.1\examples\apps'
-///
-/// # Note
-/// [`get_real_path_from_external`] is just a method for calculating a path, it doesn't check whether a path exists.
-pub(crate) fn get_real_path_from_external(
-    pkg_name: &str,
-    pkgpath: &str,
-    current_pkg_path: PathBuf,
-) -> PathBuf {
-    let mut real_path = PathBuf::new();
-    let pkg_root = fetch_metadata(current_pkg_path)
-        .map(|metadata| {
-            metadata
-                .packages
-                .get(pkg_name)
-                .map_or(PathBuf::new(), |pkg| pkg.manifest_path.clone())
-        })
-        .unwrap_or_else(|_| PathBuf::new());
-    real_path = real_path.join(pkg_root);
-
-    let pkgpath = match rm_external_pkg_name(pkgpath) {
-        Ok(path) => path,
-        Err(_) => String::new(),
-    };
-    pkgpath.split('.').for_each(|s| real_path.push(s));
-    real_path
-}
-
-#[allow(unused)]
-pub(crate) fn build_word_index_for_source_codes(
-    source_codes: HashMap<String, String>,
-    prune: bool,
-) -> anyhow::Result<HashMap<String, Vec<VirtualLocation>>> {
-    let mut index: HashMap<String, Vec<VirtualLocation>> = HashMap::new();
-    for (filepath, content) in source_codes.iter() {
-        for (key, values) in build_word_index_for_file_content_with_vfs(
-            content.to_string(),
-            filepath.to_string(),
-            prune,
-        ) {
-            index.entry(key).or_default().extend(values);
-        }
-    }
-    Ok(index)
-}
-
-pub(crate) fn build_word_index_for_file_paths(
-    paths: &[String],
-    prune: bool,
-) -> anyhow::Result<HashMap<String, Vec<Location>>> {
-    let mut index: HashMap<String, Vec<Location>> = HashMap::new();
-    for p in paths {
-        // str path to url
-        if let Ok(url) = Url::from_file_path(p) {
-            // read file content and save the word to word index
-            let text = read_file(p)?;
-            for (key, values) in build_word_index_for_file_content(text, &url, prune) {
-                index.entry(key).or_default().extend(values);
-            }
-        }
-    }
-    Ok(index)
-}
-
-/// scan and build a word -> Locations index map
-pub(crate) fn build_word_index(
-    path: String,
-    prune: bool,
-) -> anyhow::Result<HashMap<String, Vec<Location>>> {
-    let files = get_kcl_files(path.clone(), true)?;
-    build_word_index_for_file_paths(&files, prune)
-}
-
-#[allow(unused)]
-pub(crate) struct VirtualLocation {
-    pub(crate) filepath: String,
-    pub(crate) range: Range,
-}
-
-pub(crate) fn build_word_index_for_file_content_with_vfs(
-    content: String,
-    filepath: String,
-    prune: bool,
-) -> HashMap<String, Vec<VirtualLocation>> {
-    let mut index: HashMap<String, Vec<VirtualLocation>> = HashMap::new();
-    let lines: Vec<&str> = content.lines().collect();
-    let mut in_docstring = false;
-    for (li, line) in lines.into_iter().enumerate() {
-        if prune && !in_docstring && line.trim_start().starts_with("\"\"\"") {
-            in_docstring = true;
-            continue;
-        }
-        if prune && in_docstring {
-            if line.trim_end().ends_with("\"\"\"") {
-                in_docstring = false;
-            }
-            continue;
-        }
-        let words = line_to_words(line.to_string(), prune);
-        for (key, values) in words {
-            index
-                .entry(key)
-                .or_default()
-                .extend(values.iter().map(|w| VirtualLocation {
-                    filepath: filepath.clone(),
-                    range: Range {
-                        start: Position::new(li as u32, w.start_col),
-                        end: Position::new(li as u32, w.end_col),
-                    },
-                }));
-        }
-    }
-    index
-}
-
-pub(crate) fn build_word_index_for_file_content(
-    content: String,
-    url: &Url,
-    prune: bool,
-) -> HashMap<String, Vec<Location>> {
-    let mut index: HashMap<String, Vec<Location>> = HashMap::new();
-    let lines: Vec<&str> = content.lines().collect();
-    let mut in_docstring = false;
-    for (li, line) in lines.into_iter().enumerate() {
-        if prune && !in_docstring && line.trim_start().starts_with("\"\"\"") {
-            in_docstring = true;
-            continue;
-        }
-        if prune && in_docstring {
-            if line.trim_end().ends_with("\"\"\"") {
-                in_docstring = false;
-            }
-            continue;
-        }
-        let words = line_to_words(line.to_string(), prune);
-        for (key, values) in words {
-            index
-                .entry(key)
-                .or_default()
-                .extend(values.iter().map(|w| Location {
-                    uri: url.clone(),
-                    range: Range {
-                        start: Position::new(li as u32, w.start_col),
-                        end: Position::new(li as u32, w.end_col),
-                    },
-                }));
-        }
-    }
-    index
-}
-
-pub(crate) fn word_index_add(
-    from: &mut HashMap<String, Vec<Location>>,
-    add: HashMap<String, Vec<Location>>,
-) {
-    for (key, value) in add {
-        from.entry(key).or_default().extend(value);
-    }
-}
-
-pub(crate) fn word_index_subtract(
-    from: &mut HashMap<String, Vec<Location>>,
-    remove: HashMap<String, Vec<Location>>,
-) {
-    for (key, value) in remove {
-        for v in value {
-            from.entry(key.clone()).and_modify(|locations| {
-                locations.retain(|loc| loc != &v);
-            });
-        }
-    }
-}
-
-// Word describes an arbitrary word in a certain line including
-// start position, end position and the word itself.
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub struct Word {
-    start_col: u32,
-    end_col: u32,
-    word: String,
-}
-
-impl Word {
-    fn new(start_col: u32, end_col: u32, word: String) -> Self {
-        Self {
-            start_col,
-            end_col,
-            word,
-        }
-    }
-}
-
-pub fn read_file(path: &String) -> anyhow::Result<String> {
-    let text = std::fs::read_to_string(path)?;
-    Ok(text)
-}
-
-// Split one line into identifier words.
-fn line_to_words(text: String, prune: bool) -> HashMap<String, Vec<Word>> {
-    let mut result = HashMap::new();
-    let mut chars: Vec<char> = text.chars().collect();
-    chars.push('\n');
-    let mut start_pos = usize::MAX;
-    let mut continue_pos = usize::MAX - 1; // avoid overflow
-    let mut prev_word = false;
-    let mut words: Vec<Word> = vec![];
-    for (i, ch) in chars.iter().enumerate() {
-        if prune && *ch == '#' {
-            break;
-        }
-        let is_id_start = rustc_lexer::is_id_start(*ch);
-        let is_id_continue = rustc_lexer::is_id_continue(*ch);
-        // If the character is valid identifier start and the previous character is not valid identifier continue, mark the start position.
-        if is_id_start && !prev_word {
-            start_pos = i;
-        }
-        if is_id_continue {
-            // Continue searching for the end position.
-            if start_pos != usize::MAX {
-                continue_pos = i;
-            }
-        } else {
-            // Find out the end position.
-            if continue_pos + 1 == i {
-                let word = chars[start_pos..i].iter().collect::<String>().clone();
-                // skip word if it should be pruned
-                if !prune || !reserved::is_reserved_word(&word) {
-                    words.push(Word::new(start_pos as u32, i as u32, word));
-                }
-            }
-            // Reset the start position.
-            start_pos = usize::MAX;
-        }
-        prev_word = is_id_continue;
-    }
-
-    for w in words {
-        result.entry(w.word.clone()).or_insert(Vec::new()).push(w);
-    }
-    result
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        build_word_index, build_word_index_for_file_content, line_to_words, word_index_add,
-        word_index_subtract, Word,
-    };
-    use lsp_types::{Location, Position, Range, Url};
-    use std::collections::HashMap;
-    use std::path::PathBuf;
-    #[test]
-    fn test_build_word_index() {
-        let root = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-        let mut path = root.clone();
-        path.push("src/test_data/find_refs_test/main.k");
-
-        let url = lsp_types::Url::from_file_path(path.clone()).unwrap();
-        let path = path.to_str().unwrap();
-        let expect: HashMap<String, Vec<Location>> = vec![
-            (
-                "a".to_string(),
-                vec![
-                    Location {
-                        uri: url.clone(),
-                        range: Range {
-                            start: Position::new(0, 0),
-                            end: Position::new(0, 1),
-                        },
-                    },
-                    Location {
-                        uri: url.clone(),
-                        range: Range {
-                            start: Position::new(1, 4),
-                            end: Position::new(1, 5),
-                        },
-                    },
-                    Location {
-                        uri: url.clone(),
-                        range: Range {
-                            start: Position::new(2, 4),
-                            end: Position::new(2, 5),
-                        },
-                    },
-                    Location {
-                        uri: url.clone(),
-                        range: Range {
-                            start: Position::new(12, 14),
-                            end: Position::new(12, 15),
-                        },
-                    },
-                ],
-            ),
-            (
-                "c".to_string(),
-                vec![Location {
-                    uri: url.clone(),
-                    range: Range {
-                        start: Position::new(2, 0),
-                        end: Position::new(2, 1),
-                    },
-                }],
-            ),
-            (
-                "b".to_string(),
-                vec![Location {
-                    uri: url.clone(),
-                    range: Range {
-                        start: Position::new(1, 0),
-                        end: Position::new(1, 1),
-                    },
-                }],
-            ),
-            (
-                "n".to_string(),
-                vec![
-                    Location {
-                        uri: url.clone(),
-                        range: Range {
-                            start: Position::new(8, 4),
-                            end: Position::new(8, 5),
-                        },
-                    },
-                    Location {
-                        uri: url.clone(),
-                        range: Range {
-                            start: Position::new(11, 4),
-                            end: Position::new(11, 5),
-                        },
-                    },
-                ],
-            ),
-            (
-                "b".to_string(),
-                vec![Location {
-                    uri: url.clone(),
-                    range: Range {
-                        start: Position::new(1, 0),
-                        end: Position::new(1, 1),
-                    },
-                }],
-            ),
-            (
-                "Name".to_string(),
-                vec![
-                    Location {
-                        uri: url.clone(),
-                        range: Range {
-                            start: Position::new(4, 7),
-                            end: Position::new(4, 11),
-                        },
-                    },
-                    Location {
-                        uri: url.clone(),
-                        range: Range {
-                            start: Position::new(8, 7),
-                            end: Position::new(8, 11),
-                        },
-                    },
-                    Location {
-                        uri: url.clone(),
-                        range: Range {
-                            start: Position::new(11, 7),
-                            end: Position::new(11, 11),
-                        },
-                    },
-                ],
-            ),
-            (
-                "name".to_string(),
-                vec![
-                    Location {
-                        uri: url.clone(),
-                        range: Range {
-                            start: Position::new(5, 4),
-                            end: Position::new(5, 8),
-                        },
-                    },
-                    Location {
-                        uri: url.clone(),
-                        range: Range {
-                            start: Position::new(12, 8),
-                            end: Position::new(12, 12),
-                        },
-                    },
-                ],
-            ),
-            (
-                "demo".to_string(),
-                vec![Location {
-                    uri: url.clone(),
-                    range: Range {
-                        start: Position::new(0, 5),
-                        end: Position::new(0, 9),
-                    },
-                }],
-            ),
-            (
-                "Person".to_string(),
-                vec![
-                    Location {
-                        uri: url.clone(),
-                        range: Range {
-                            start: Position::new(7, 7),
-                            end: Position::new(7, 13),
-                        },
-                    },
-                    Location {
-                        uri: url.clone(),
-                        range: Range {
-                            start: Position::new(10, 5),
-                            end: Position::new(10, 11),
-                        },
-                    },
-                ],
-            ),
-            (
-                "p2".to_string(),
-                vec![Location {
-                    uri: url.clone(),
-                    range: Range {
-                        start: Position::new(10, 0),
-                        end: Position::new(10, 2),
-                    },
-                }],
-            ),
-        ]
-        .into_iter()
-        .collect();
-        match build_word_index(path.to_string(), true) {
-            Ok(actual) => {
-                assert_eq!(expect, actual)
-            }
-            Err(_) => assert!(false, "build word index failed. expect: {:?}", expect),
-        }
-    }
-
-    #[test]
-    fn test_word_index_add() {
-        let loc1 = Location {
-            uri: Url::parse("file:///path/to/file.k").unwrap(),
-            range: Range {
-                start: Position::new(0, 0),
-                end: Position::new(0, 4),
-            },
-        };
-        let loc2 = Location {
-            uri: Url::parse("file:///path/to/file.k").unwrap(),
-            range: Range {
-                start: Position::new(1, 0),
-                end: Position::new(1, 4),
-            },
-        };
-        let mut from = HashMap::from([("name".to_string(), vec![loc1.clone()])]);
-        let add = HashMap::from([("name".to_string(), vec![loc2.clone()])]);
-        word_index_add(&mut from, add);
-        assert_eq!(
-            from,
-            HashMap::from([("name".to_string(), vec![loc1.clone(), loc2.clone()],)])
-        );
-    }
-
-    #[test]
-    fn test_word_index_subtract() {
-        let loc1 = Location {
-            uri: Url::parse("file:///path/to/file.k").unwrap(),
-            range: Range {
-                start: Position::new(0, 0),
-                end: Position::new(0, 4),
-            },
-        };
-        let loc2 = Location {
-            uri: Url::parse("file:///path/to/file.k").unwrap(),
-            range: Range {
-                start: Position::new(1, 0),
-                end: Position::new(1, 4),
-            },
-        };
-        let mut from = HashMap::from([("name".to_string(), vec![loc1.clone(), loc2.clone()])]);
-        let remove = HashMap::from([("name".to_string(), vec![loc2.clone()])]);
-        word_index_subtract(&mut from, remove);
-        assert_eq!(
-            from,
-            HashMap::from([("name".to_string(), vec![loc1.clone()],)])
-        );
-    }
-
-    #[test]
-    fn test_line_to_words() {
-        let lines = [
-            "schema Person:",
-            "name. name again",
-            "some_word word !word",
-            "# this line is a single-line comment",
-            "name # end of line comment",
-        ];
-
-        let expects: Vec<HashMap<String, Vec<Word>>> = vec![
-            vec![(
-                "Person".to_string(),
-                vec![Word {
-                    start_col: 7,
-                    end_col: 13,
-                    word: "Person".to_string(),
-                }],
-            )]
-            .into_iter()
-            .collect(),
-            vec![
-                (
-                    "name".to_string(),
-                    vec![
-                        Word {
-                            start_col: 0,
-                            end_col: 4,
-                            word: "name".to_string(),
-                        },
-                        Word {
-                            start_col: 6,
-                            end_col: 10,
-                            word: "name".to_string(),
-                        },
-                    ],
-                ),
-                (
-                    "again".to_string(),
-                    vec![Word {
-                        start_col: 11,
-                        end_col: 16,
-                        word: "again".to_string(),
-                    }],
-                ),
-            ]
-            .into_iter()
-            .collect(),
-            vec![
-                (
-                    "some_word".to_string(),
-                    vec![Word {
-                        start_col: 0,
-                        end_col: 9,
-                        word: "some_word".to_string(),
-                    }],
-                ),
-                (
-                    "word".to_string(),
-                    vec![
-                        Word {
-                            start_col: 10,
-                            end_col: 14,
-                            word: "word".to_string(),
-                        },
-                        Word {
-                            start_col: 16,
-                            end_col: 20,
-                            word: "word".to_string(),
-                        },
-                    ],
-                ),
-            ]
-            .into_iter()
-            .collect(),
-            HashMap::new(),
-            vec![(
-                "name".to_string(),
-                vec![Word {
-                    start_col: 0,
-                    end_col: 4,
-                    word: "name".to_string(),
-                }],
-            )]
-            .into_iter()
-            .collect(),
-        ];
-        for i in 0..lines.len() {
-            let got = line_to_words(lines[i].to_string(), true);
-            assert_eq!(expects[i], got)
-        }
-    }
-
-    #[test]
-    fn test_build_word_index_for_file_content() {
-        let content = r#"schema Person:
-    """
-    This is a docstring.
-    Person is a schema which defines a person's name and age.
-    """
-    name: str # name must not be empty
-    # age is a positive integer
-    age: int
-"#;
-        let mock_url = Url::parse("file:///path/to/file.k").unwrap();
-        let expects: HashMap<String, Vec<Location>> = vec![
-            (
-                "Person".to_string(),
-                vec![Location {
-                    uri: mock_url.clone(),
-                    range: Range {
-                        start: Position::new(0, 7),
-                        end: Position::new(0, 13),
-                    },
-                }],
-            ),
-            (
-                "name".to_string(),
-                vec![Location {
-                    uri: mock_url.clone(),
-                    range: Range {
-                        start: Position::new(5, 4),
-                        end: Position::new(5, 8),
-                    },
-                }],
-            ),
-            (
-                "age".to_string(),
-                vec![Location {
-                    uri: mock_url.clone(),
-                    range: Range {
-                        start: Position::new(7, 4),
-                        end: Position::new(7, 7),
-                    },
-                }],
-            ),
-        ]
-        .into_iter()
-        .collect();
-
-        let got = build_word_index_for_file_content(content.to_string(), &mock_url.clone(), true);
-        assert_eq!(expects, got)
     }
 }
